Added a heap monitoring class and use for gbst-test program
This commit adds a new class, heap_mon_class to provide a method for
tracking the heap process utilization during the running of the program.  It
has the capability of taking a heap snapshot at any time required and
keeping track of the minimum, maximum, current, and minute average for a
number of characteristics that are returned from the system malloc.h
mallinfo2() function.  This information can then be displayed either in the
ncurses display screen if that is in use, or as output to the standard
terminal scroll.

Add the necessary instructions to implement the heap monitoring capability
to the gbtree test program, and provides a number of observation points as
well as a continual heap status monitoring display during the testing
process.  The display is currently implemented for the ncurses run
environment in the graph window area as that is not yet being used.

The doc/repository-notes.txt file captures information about development
issues that came up during the build up of the various parts of the file
organizer application.  The notes also provide some small insight into the
usefulness of George's binary search tree modules.