Changes to the heap monitor min and max display and check version
The display of minimum and maximum originally showed the absolute minimums
and maximums that were seen by the heap monitor, but I realized that as the
program progresses, it is more interesting to see the recent trends for
those values, and changed the heap_mon_class to show what the minimums and
maximums have been over the previous minute with reset updates five times
per minute.  So, five times per minute, the previous minimum or maximum is
replaced with the current previous minute information if that value is
closer to the current value than what was there before.

Instead of maintaining a record of the steady clock time of the previous
heap sampling, I now just keep track of the previous run_µsec instead and
compute the interval from the difference between the run_µsec and the
previous one.

A GNU lib C version check is needed to insure that the recently added
capability provided by the mallinfo2 function is available.  The version
check is used to force the compiled program to revert to the older mallinfo
function if mallinfo2 is not available.