Updated heap_mon_class::get_info() method process efficiency
Using a newly discovered struct member access method, I have been able to
significantly improve the efficiency of the get-info() method.  Instead of
having to pass references to every member needed by the calc subfunction, and
creating numerous temporary arrays, by using the capability to pass pointers to
member for the required structs, the call is simplified and the subfunction
process is more straightforward and easier to understand.

I wondered initially if this was possible, but I misunderstood the information
about the pointer to member and thought it was a fixed pointer to a specific
instantiation instead of a generic pointer to member for the type.  Maybe
someday, I will learn enough about C++ that I can avoid these types of gaffs.