Update ncursio heap monitor update method
Instead of using the manage_debug_win() method for doing the
heap monitor display, a new method has been implemented
exclusively for that task.  The new method is called from the
ncursio main loop whenever it detects that new information is
available.  This provides enhanced visualization of the heap
monitoring status when the ncursio screen is active.