Streamline heap and stream monitoring utilization
Instead of passing a reference to an output stream to these methods, I now
let them declare their own local stringstream and then return the string
that is produced.  This allows me to simplify the usage of the methods since
the returned string can be used directly by whatever output method is
needed, which usually results in reducing the code needed from three lines
to just one.

Also, simplified the use of the base search variable info stream (bsvistrm)
by making it a general stringstream instead of just an output stream.  This
permits it to be used as input for displaying the ncurses info window
information instead of getting successive substrings from the ostringstream
as was done in the past.  The previous method resulted in large heap
allocations whenever a substring was needed, that are no longer required
with the change.  However, it is now required for the main program to make
sure that the input stream pointer is set to the beginning of the file
before doing the file output of the bsvistrm record.