Fixed library problem with the build script
The bld-gbst build script did not check for the existance of the lib folder,
and if it is not there, the library would not be built.  This commit insures
that the lib folder exists and that any old libbtutil.a library files are
removed before starting the build which will create a new copy of the
directory.