Review of the comments in the gbtree folder source files
As a result of the review of comments in the gbtree folder source files, a number
of changes have been made to align the comments with the actual current operation
of the code.  All of the files except the utf8-name-store and the utf8-rcrd-type
modules have been reviewed.  The remaining two modules are to be the subject of an
upcoming code modification, and thus the review of their comments will be a part
of the modification.