Update retrieval of gbtree ID values
The three gbtree ID values, btree_parent, btree_child_left, and
btree_child_right, that are contained in the gbtree record are stored in an
unsigned bitfield, but must be interpreted as an unbalanced integer.  This
commit defines and documents a private inline gbtree method that does the
appropriate decoding of the values, and that method is then used by the
public methods defined to retrieve the relevant ID value.