Initial commit for the macro redesign
This commit provides the first of the new macros that will comprise the
macro re-design for the gbst-test project.  The existing macro usage
scenario was developed as a result of my reluctance to fully study the
capabilities of the cpp preprocessor, and since then, I have realized that a
much more efficient solution is available, and will be implemented initially
for the macros needed for debugging.